Working SQS refactor
