+++ ecs_composex/common/__init__.py
@@ -203,14 +203,13 @@ def setup_logging():
     level = environ.get("LOGLEVEL")
     default_level = True
     formats = {
-        'INFO': logthings.Formatter(
-            "%(asctime)s [%(levelname)s], %(message)s",
-            "%Y-%m-%d %H:%M:%S",
+        "INFO": logthings.Formatter(
+            "%(asctime)s [%(levelname)s], %(message)s", "%Y-%m-%d %H:%M:%S",
         ),
-        'DEBUG': logthings.Formatter(
+        "DEBUG": logthings.Formatter(
             "%(asctime)s [%(levelname)s], %(filename)s.%(lineno)d , %(funcName)s, %(message)s",
             "%Y-%m-%d %H:%M:%S",
-        )
+        ),
     }
 
     if level is not None and isinstance(level, str):
@@ -227,11 +226,11 @@ def setup_logging():
 
     if not the_logger.handlers:
         if default_level:
-            formatter = formats['INFO']
+            formatter = formats["INFO"]
         elif KEYISSET(level.upper(), formats):
             formatter = formats[level.upper()]
         else:
-            formatter = formats['DEBUG']
+            formatter = formats["DEBUG"]
         handler = logthings.StreamHandler(sys.stdout)
         handler.setFormatter(formatter)
         the_logger.addHandler(handler)
