+++ ecs_composex/ecs/ecs_template.py
@@ -2,11 +2,9 @@
 """ Core ECS Template building """
 
 from troposphere import Ref, Sub, Tags
-from troposphere.cloudformation import Stack
 from troposphere.ec2 import SecurityGroup
 from troposphere.logs import LogGroup
 
-from ecs_composex.ecs_composex import generate_x_resource_configs
 from ecs_composex.common import LOG
 from ecs_composex.common import cfn_params, build_template, KEYISSET, add_parameters
 from ecs_composex.common.cfn_params import ROOT_STACK_NAME_T, ROOT_STACK_NAME
@@ -77,14 +75,14 @@ def add_services_stacks(
     :param kwargs: optional arguments
     :type kwargs: dicts or set
     """
-    x_resources_configs = generate_x_resource_configs(compose_content, **kwargs)
+    # x_resources_configs = generate_x_resource_configs(compose_content, **kwargs)
     for service_name in compose_content[ecs_params.RES_KEY]:
         service = compose_content[ecs_params.RES_KEY][service_name]
         service_set = generate_service_template(
             compose_content,
             service_name,
             service,
-            x_resources_configs,
+            # x_resources_configs,
             tags=tags,
             session=session,
             **kwargs,
@@ -101,13 +99,15 @@ def add_services_stacks(
         LOG.debug(dependencies)
         parameters.update(service_set[1])
         if service_set[0]:
-            root_tpl.add_resource(ComposeXStack(
-                service_name,
-                template=service_set[0],
-                Parameters=parameters,
-                DependsOn=dependencies,
-                **kwargs
-            ))
+            root_tpl.add_resource(
+                ComposeXStack(
+                    service_name,
+                    template=service_set[0],
+                    Parameters=parameters,
+                    DependsOn=dependencies,
+                    **kwargs,
+                )
+            )
         else:
             Warning(
                 f"Template for service {service_name}" "was not successfully generated"
@@ -126,7 +126,7 @@ def generate_services_templates(compose_content, session=None, **kwargs):
         vpc_params.VPC_ID,
         vpc_params.PUBLIC_SUBNETS,
         vpc_params.APP_SUBNETS,
-        vpc_params.VPC_MAP_ID
+        vpc_params.VPC_MAP_ID,
     ]
     template = build_template("Root template for ECS Services", parameters)
     if tags_params:
